--- src/main/java/com/airdnb/clone/domain/filter/repository/ReservationQuerydsl.java
@@ -1,9 +1,0 @@
-package com.airdnb.clone.domain.filter.repository;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public interface ReservationQuerydsl {
-
-    List<Long> findUnavailableStayIdsByCheckInOut(LocalDate checkInDate, LocalDate checkOutDate);
-}
