+++ src/main/java/com/airdnb/clone/domain/reservation/repository/ReservationRepository.java
@@ -1,6 +1,5 @@
-package com.airdnb.clone.domain.reservation;
+package com.airdnb.clone.domain.reservation.repository;
 
-import com.airdnb.clone.domain.filter.repository.ReservationQuerydsl;
 import com.airdnb.clone.domain.reservation.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
