+++ src/main/java/com/airdnb/clone/domain/stay/repository/StayRepository.java
@@ -1,10 +1,9 @@
 package com.airdnb.clone.domain.stay.repository;
 
-import com.airdnb.clone.domain.filter.repository.StayQuerydsl;
 import com.airdnb.clone.domain.stay.entity.Stay;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StayRepository extends JpaRepository<Stay, Long>, StayQuerydsl {
+public interface StayRepository extends JpaRepository<Stay, Long> {
 }
